+++ sdk/eventhub/Azure.Messaging.EventHubs/api/Azure.Messaging.EventHubs.netstandard2.0.cs
@@ -230,8 +230,10 @@ namespace Azure.Messaging.EventHubs.Consumer
     public partial class ReadEventOptions
     {
         public ReadEventOptions() { }
+        public int CacheEventCount { get { throw null; } set { } }
         public System.TimeSpan? MaximumWaitTime { get { throw null; } set { } }
         public long? OwnerLevel { get { throw null; } set { } }
+        public int PrefetchCount { get { throw null; } set { } }
         public bool TrackLastEnqueuedEventProperties { get { throw null; } set { } }
         [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
         public override bool Equals(object obj) { throw null; }
@@ -258,6 +260,7 @@ namespace Azure.Messaging.EventHubs.Primitives
         public Azure.Messaging.EventHubs.EventHubConnectionOptions ConnectionOptions { get { throw null; } set { } }
         public Azure.Messaging.EventHubs.Consumer.EventPosition DefaultStartingPosition { get { throw null; } set { } }
         public string Identifier { get { throw null; } set { } }
+        public Azure.Messaging.EventHubs.Processor.LoadBalancingStrategy LoadBalancingStrategy { get { throw null; } set { } }
         public System.TimeSpan LoadBalancingUpdateInterval { get { throw null; } set { } }
         public System.TimeSpan? MaximumWaitTime { get { throw null; } set { } }
         public System.TimeSpan PartitionOwnershipExpirationInterval { get { throw null; } set { } }
@@ -364,6 +367,11 @@ namespace Azure.Messaging.EventHubs.Primitives
 }
 namespace Azure.Messaging.EventHubs.Processor
 {
+    public enum LoadBalancingStrategy
+    {
+        Balanced = 0,
+        Greedy = 1,
+    }
     public partial class PartitionClosingEventArgs
     {
         public PartitionClosingEventArgs(string partitionId, Azure.Messaging.EventHubs.Processor.ProcessingStoppedReason reason, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { }